move deleting of pqsdata to segmetarw
